Rename external api header
