+++ library/include/ck/library/tensor_operation_instance/gpu/normalization_fwd_swish.hpp
